+++ test/tests/braintree-manager.test.ts
@@ -2,26 +2,18 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { expect } from '@open-wc/testing';
 
-import { BraintreeManager } from '../src/braintree-manager/braintree-manager';
-import { HostingEnvironment } from '../src/braintree-manager/braintree-interfaces';
-import { MockPaymentClients } from './mocks/mock-payment-clients';
-import { MockEndpointManager } from './mocks/mock-endpoint-manager';
-import { mockHostedFieldConfig } from './mocks/mock-hosted-fields-config';
-import { PromisedSingleton } from '@internetarchive/promised-singleton';
-import { DonationRequest } from '../src/models/request-models/donation-request';
-import { DonationResponse } from '../src/models/response-models/donation-response';
-import { ErrorResponse } from '../src/models/response-models/error-models/error-response';
-import { PaymentProvider } from '../src/models/common/payment-provider-name';
-import { DonationPaymentInfo } from '../src/models/donation-info/donation-payment-info';
-import { DonationType } from '../src/models/donation-info/donation-type';
-import { BillingInfo } from '../src/models/common/billing-info';
-import { MockBillingInfo } from './mocks/mock-billing-info';
-import { MockCustomerInfo } from './mocks/mock-customer-info';
-import { callback } from 'braintree-web';
-import { Configuration } from 'braintree-web/modules/client';
-import { MockDeviceDataCollector } from './mocks/payment-clients/mock-data-collector';
-import { SuccessResponse } from '../src/models/response-models/success-models/success-response';
-import { mockSuccessResponse } from './mocks/mock-success-response';
+import { BraintreeManager } from '../../src/braintree-manager/braintree-manager';
+import { HostingEnvironment } from '../../src/braintree-manager/braintree-interfaces';
+import { MockPaymentClients } from '../mocks/mock-payment-clients';
+import { MockEndpointManager } from '../mocks/mock-endpoint-manager';
+import { mockHostedFieldConfig } from '../mocks/mock-hosted-fields-config';
+import { PaymentProvider } from '../../src/models/common/payment-provider-name';
+import { DonationPaymentInfo } from '../../src/models/donation-info/donation-payment-info';
+import { DonationType } from '../../src/models/donation-info/donation-type';
+import { MockBillingInfo } from '../mocks/mock-billing-info';
+import { MockCustomerInfo } from '../mocks/mock-customer-info';
+import { MockDeviceDataCollector } from '../mocks/payment-clients/mock-data-collector';
+import { mockSuccessResponse } from '../mocks/mock-success-response';
 
 describe('Braintree Manager', () => {
   it('can be initialized', async () => {
